refactor: extract admin authentication to shared concern
