+++ app/controllers/api/v1/admin/orders_controller.rb
@@ -1,4 +1,6 @@
 class Api::V1::Admin::OrdersController < Api::V1::ApplicationController
+  include AdminAuthenticatable
+  
   before_action :require_admin
   before_action :set_order, only: [ :show, :update ]
 
@@ -88,14 +90,6 @@ class Api::V1::Admin::OrdersController < Api::V1::ApplicationController
   end
 
   private
-
-  def require_admin
-    unless current_user.admin?
-      render json: {
-        status: { message: "Access denied. Admin privileges required." }
-      }, status: :forbidden
-    end
-  end
 
   def set_order
     @order = Order.find(params[:id])
